// integra-ifood/src/types/print.types.ts
export type EnderecoMascara = {
    Endereco: string;
    mascara: string;
};


export type  EtiquetaCabo = {
    nunota: number;
    parceiro: string;
    vendedor: string;
    codprod: number;
    descrprod: string;
    qtdneg: number;
    codbarras: string;
};

export type Localizacoes = {
    Rua: string;
    Predio: string;
    Nivel: string;
    Apartamento: string;
    Endereco: string;
    Armazenamento: string;
}